✨ 1.21.8

+++ src/main/java/com/tiviacz/travelersbackpack/util/Reference.java
@@ -11,9 +11,9 @@ public class Reference {
     public static final int POTION = 250;
 
     //Screen IDs
-    public static final int ITEM_SCREEN_ID = 1;
-    public static final int WEARABLE_SCREEN_ID = 2;
-    public static final int BLOCK_ENTITY_SCREEN_ID = 3;
+    public static final byte ITEM_SCREEN_ID = 1;
+    public static final byte WEARABLE_SCREEN_ID = 2;
+    public static final byte BLOCK_ENTITY_SCREEN_ID = 3;
 
     //Translation Keys
     public static final String NO_SPACE = "action.travelersbackpack.unequip_nospace";
